# docs/swagger.yaml
basePath: /api
definitions:
...
  models.ResponseMessage:
    properties:
      content: {}
      message:
        example: Action completed successfully
        type: string
      result:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API per gestionar operadors i workcenters
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Actions Service API
  version: "1.0"
paths:
...
  /ws/workcenter/{id}:
    get:
      consumes:
      - application/json
      description: Estableix una connexió WebSocket per rebre actualitzacions específiques
        d'un workcenter
      parameters:
      - description: ID del Workcenter
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "404":
          description: Workcenter no trobat
          schema:
            $ref: '#/definitions/models.ResponseMessage'
      summary: WebSocket de connexió per workcenter
      tags:
      - websocket
schemes:
- http
- https
swagger: "2.0"